+++ src/eps/tests/test9.c
@@ -301,13 +301,13 @@ PetscErrorCode MyEigenSort(PetscScalar ar,PetscScalar ai,PetscScalar br,PetscSca
       output_file: output/test9_1.out
 
    test:
-      suffix: 9
+      suffix: 5_ksphpddm
       args: -eps_nev 3 -st_type sinvert -eps_target .5 -st_ksp_type hpddm -st_ksp_hpddm_type gcrodr -eps_ncv 10
       requires: hpddm
       output_file: output/test9_5.out
 
    test:
-      suffix: 10
+      suffix: 5_pchpddm
       args: -eps_nev 3 -st_type sinvert -eps_target .5 -st_pc_type hpddm -st_pc_hpddm_coarse_pc_type lu -eps_ncv 10
       requires: hpddm
       output_file: output/test9_5.out